+++ app/Modules/Discount/Entity/EnabledDiscountAbstract.php
@@ -5,7 +5,7 @@ namespace App\Modules\Discount\Entity;
 
 abstract class EnabledDiscountAbstract
 {
-    abstract public static function isEnabled(Discount $discount, int $cost = null): bool;
+    abstract public static function isEnabled(Discount $discount, float $cost = null): bool;
     abstract public static function name(): string;
     abstract public static function caption(string $from_to): string;
     abstract public static function widget(): string;
